moved findEmailForInitiatorRole to delegate
